Use a u32 to represent ReplicaId
This will avoid using a BigInt in the js binding.

+++ memo_core/src/lib.rs
@@ -23,7 +23,7 @@ use std::borrow::Cow;
 use std::fmt;
 use std::io;
 
-pub type ReplicaId = u64;
+pub type ReplicaId = u32;
 pub type UserId = u64;
 pub type Oid = [u8; 20];
